+++ LevelUp/Domain/Goals/GoalStatus.cs
@@ -1,8 +1,0 @@
-namespace LevelUp.Domain.Goals;
-
-public enum GoalStatus
-{
-    Active,
-    Completed,
-    Archived
-}